работа с темами перенесена в отдельный класс
